push

+++ install.sh
@@ -1,7 +1,28 @@
 #!/bin/bash
 
-sudo apt-get install git ffmpeg libsm6 libxext6 curl -y
+# Check we're in the home directory
+if [ ! -d "./home" ]; then
+  read -p "Are you sure you want to run this outside of the home directory? [y/N] " -n 1 -r
+  if [[ ! $REPLY =~ ^[Yy]$ ]]; then
+    exit 1
+  fi
+fi
+
+# Remove existing repo
 rm -rf third-eye-api
+
+# Dependencies for everything
+sudo apt-get install git ffmpeg libsm6 libxext6 curl -y
+
+# Install nvm, so we can install `degit` to quickly clone git repos
+curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.39.1/install.sh | bash
+export NVM_DIR="$HOME/.nvm"
+[ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"                   # This loads nvm
+[ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion" # This loads nvm bash_completion
+nvm install 16
+npm i -g degit
+
+# The only git repo where we're interested in the git history
 git clone --depth 1 "https://github.com/andyshand/third-eye-api.git"
 
 source /root/venv/bin/activate
